+++ src/Components/gallery.js
@@ -3,7 +3,6 @@
 //Components
 import NavBar from "./navbar";
 import Preloader from "./preloader";
-import ScrollButton from "./scroll-to-top";
 
 //Icons
 import icon from "../Icons/puzzle.png";
@@ -38,7 +37,6 @@ class Gallery extends PureComponent {
   render() {
     const { width } = this.state;
     const { isLoading, data, url, copy, isOpen, layout, showModal } = this.props;
-    console.log("TCL: Gallery -> render -> data", data)
 
     const parentClassChange = layout ? "mosaic" : "container-fluid galleryContainer";
 
@@ -50,7 +48,7 @@ class Gallery extends PureComponent {
             data.map(image => {
               return (
                   <Suspense key={image.sys.id} fallback={<Preloader />}>
-                      <GalleryImage images={image.fields.images} layout={layout} showModal={showModal} parentTheme={image.fields.theme + "-"} />
+                    <GalleryImage images={image.fields.images} layout={layout} showModal={showModal} parentTheme={image.fields.theme + "-"} />
                   </Suspense>
               );
             })
@@ -60,13 +58,6 @@ class Gallery extends PureComponent {
           {width > 667 ? <img src={icon} alt="icons8.com" className="layout-change-icon" onClick={this.handleCssChange} /> : null}
           {isLoading && <Preloader />}
         </div>
-        <footer>
-          <ScrollButton />
-          <p>
-            &copy; Copyright message goes Here
-          </p>
-        </footer>
-
       </Fragment>
     )
   }